Give the cash bag the 2.5x respawn delay it was written for
CASH_BAG is 25 and CASH_ROLL is 10, so the ratio the branch exists to apply
is 2.5 -- but both are ints, and the quotient truncated to 2 before it ever
reached the float multiply. The explicit (float) cast on the cvar in the
same expression shows the author was thinking in floats; the parenthesised
constant ratio beside it was missed.

Bags have therefore been respawning 20% sooner than designed since the mode
was written. Fixing it slows them down, so this is a live balance change on
the Bagman rotation, not just a correctness fix.

+++ src/g_teamplay.c
@@ -219,7 +219,7 @@ void SP_dm_cashspawn( edict_t *self )
 	}
 	else	// bag, so longer delay
 	{
-		self->delay = (float)g_cashspawndelay->value * (CASH_BAG / CASH_ROLL);
+		self->delay = (float)g_cashspawndelay->value * ((float)CASH_BAG / CASH_ROLL);
 	}
 
 	if (!self->speed)
